avahi: (user) don't conflict with self #309
If during registration of own records there is an address change or
address removed and re-added then avahi-daemon interprets its own probe
as conflicting (due to presence of different or temporary non-existing
address in the probe). And then it changes its name to "hostname-2" and
retries.
Change of hostname without good reason could be unexpected and annoying.
To resolve this we ignore presence of non-matching records if they were
sent by us.