qos - use 'gum' bit for default and multicast
Aligns with upstream code - avoid repeated scans of the table for
default-handled packets. Prior to this only user rules without
transferred limits would stick and avoid future scans.

One effect is that use of the "transferred" field must cover all ranges
from zero, otherwise the default rule will stick and the class won't
change as the count changes.

Eg, this is valid:

   any 0-128 high
   any 128-1536 medium

as the table will keep being scanned up to 1536K. After 1536K, the default
rule will hit and stick, saving further scans.

This is not valid:

   port 6881:6889 any 1536- lowest

as the connection will stick on default (low) before the 1536 is
reached. It must be:

   port 6881:6889 any 0-1536 low
   port 6881:6889 any 1536- lowest

(Potential future optimisation - make a "transferred 1536-" rule stick).