Enable QoS overhead for MIPS and fine-tune
Make the QoS overhead setting operational for MIPS, including in the
GUI.

Add new GUI profile options matching upstream.

Add support for qos_atm and qos_mpu NVRAM variables, along with the
alternative of specifying as high digits of qos_overhead for GUI.

Subtract 14 from overhead passed to htb qdisc on Ethernet interfaces, as
htb counts the Ethernet header regardless, whereas Cake, and the values
in the GUI assume the Ethernet header is counted as overhead.