QoS: eliminate use of CONNMARK --set-return
CONNMARK --set-return is a weird kernel hack that doesn't work at all
properly in newer kernels, and is questionable for old ones.

The TQoS code uses separate CONNMARK --set and RETURN rules for later
kernels, but that means two iptables rules for every entry, which is
what the CONNMARK --set-return was intended to avoid. Plus it fails to
actually set the current packet's mark.

Use a new "goto LOWEST" form that jumps to a chain that does the MARK
and CONNMARK then returns from QOSO, so we only need one iptable rule
per entry, reducing processing overhead for the rule search.

Also take care to not put the 0x100 gum flag from the ctmark into the
fwmark, meaning we can eliminate the tc fw filters checking for 0x104 as
well as 0x4.